% Auto-generated: do not edit by hand
\name{''DashRotate}

\alias{''DashRotate}

\title{DashRotate component}

\description{

}

\usage{
''DashRotate(id=NULL, delay=NULL, message=NULL, orientation=NULL,
reload=NULL, timing=NULL)
}

\arguments{
\item{id}{Character. }

\item{delay}{Numeric. }

\item{message}{A list of or a singular dash component, string or number. }

\item{orientation}{Character. }

\item{reload}{Logical. }

\item{timing}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

